Add the ability to monitor around an event.
This PR lets watchers get before and after views of
an event. The callbacks are invoked as a process enters
the context of the event. The callbacks should then yield
once to pass control back to the main flow, and each callback
will be allowed to continue when the event context exits.

Signed-off-by: Christopher Sams <csams@redhat.com>