Use hostname for group.instance.id if configured
We want to be able to use the static instance id feature of kafka so
that we can avoid unnecessary rebalances on the event of a consumer
crashing. This allows the consumer to come back up and get the same
partition it had previously so we don't pass around failed messages that
were not committed.

Signed-off-by: Stephen Adams <tsadams@gmail.com>