feat: allow multiple topic subscriptions
We need to be able to just accept a list of topics instead of assuming
we're just passing a string into a list. Since consumers can subscribe
to multiple topics, we should probably prefer a list in the yaml as
well.

To not break any apps using the library currently, I went with an if
statement, but it would probably be best to figure out how to get this
to be the default behavior moving forward.

Yaml would then look like:

    incoming_topics:
      - some.topic
      - some.other.topic

Signed-off-by: Stephen Adams <tsadams@gmail.com>