Add rebalance metrics for Kafka consumers
As rebalancing is a disruptive but necessary part of Kafka's
ConsumerGroup functionality, keeping tabs on their frequency and effect
on a Kafka consumer's operation queue size can offer extra insight into
the health of the ecosystem. This attempts to configure and enable
pertinent Prometheus metrics, though individual applications will need
to expose those metrics to be scraped.

Signed-off-by: Chris Mitchell <cmitchel@redhat.com>